# docs/openapi.yaml
openapi: 3.0.3
info:
  title: abetworks crm API
  description: The official REST API for the abetworks crm platform.
  version: 1.0.0
servers:
  - url: https://api.abetworks-crm.com/v1
    description: Production Server
  - url: http://localhost:3000/api/v1
    description: Local Development Server
tags:
  - name: Auth
    description: Authentication operations.
  - name: Accounts
    description: Operations related to customer accounts.
  - name: Contacts
    description: Operations related to contacts within accounts.
  - name: Leads
    description: Operations related to sales leads.
  - name: Opportunities
    description: Operations related to sales opportunities.
  - name: Invoices
    description: Operations related to invoices.
  - name: Tasks
    description: Operations related to tasks.

paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: User login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '401':
          description: Unauthorized

  /auth/register:
    post:
      tags:
        - Auth
      summary: User registration
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmUser'
        '409':
          description: Conflict, user already exists

  /accounts:
    get:
      tags:
        - Accounts
      summary: List all accounts
      description: Retrieves a paginated list of accounts for the user's organization.
      operationId: getAccounts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination.
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page.
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: A list of accounts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized
    post:
      tags:
        - Accounts
      summary: Create a new account
      description: Adds a new customer account to the user's organization.
      operationId: createAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountDto'
      responses:
        '201':
          description: Account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Invalid input data.
        '401':
          description: Unauthorized

  /accounts/{accountId}:
    get:
      tags:
        - Accounts
      summary: Get an account by ID
      description: Retrieves a single account by its unique ID, scoped to the user's organization.
      operationId: getAccountById
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          description: The ID of the account to retrieve.
          schema:
            type: integer
      responses:
        '200':
          description: The requested account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found.
        '401':
          description: Unauthorized

  /accounts/{accountId}/contacts:
    get:
      tags:
        - Contacts
      summary: List contacts for an account
      operationId: getContactsForAccount
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of contacts for the specified account.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
  
  /accounts/{accountId}/opportunities:
    get:
      tags:
        - Opportunities
      summary: List opportunities for an account
      operationId: getOpportunitiesForAccount
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of opportunities for the specified account.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Opportunity'

  /contacts:
    get:
      tags:
        - Contacts
      summary: List all contacts
      operationId: getContacts
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: A list of contacts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
    post:
      tags:
        - Contacts
      summary: Create a new contact
      operationId: createContact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactDto'
      responses:
        '201':
          description: Contact created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'

  /contacts/{contactId}:
    get:
      tags:
        - Contacts
      summary: Get a contact by ID
      operationId: getContactById
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested contact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          description: Contact not found.
    patch:
      tags:
        - Contacts
      summary: Update a contact
      operationId: updateContact
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactDto'
      responses:
        '200':
          description: Contact updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
    delete:
      tags:
        - Contacts
      summary: Delete a contact
      operationId: deleteContact
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Contact deleted successfully.

  /leads:
    get:
      tags:
        - Leads
      summary: List all leads
      operationId: getLeads
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of leads.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lead'
    post:
      tags:
        - Leads
      summary: Create a new lead
      operationId: createLead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadDto'
      responses:
        '201':
          description: Lead created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'

  /leads/{leadId}:
    get:
      tags:
        - Leads
      summary: Get a lead by ID
      operationId: getLeadById
      security:
        - bearerAuth: []
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested lead.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          description: Lead not found.
    patch:
      tags:
        - Leads
      summary: Update a lead
      operationId: updateLead
      security:
        - bearerAuth: []
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeadDto'
      responses:
        '200':
          description: Lead updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
    delete:
      tags:
        - Leads
      summary: Delete a lead
      operationId: deleteLead
      security:
        - bearerAuth: []
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Lead deleted successfully.
  
  /opportunities:
    get:
      tags:
        - Opportunities
      summary: List all opportunities
      operationId: getOpportunities
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of opportunities.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Opportunity'
    post:
      tags:
        - Opportunities
      summary: Create a new opportunity
      operationId: createOpportunity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOpportunityDto'
      responses:
        '201':
          description: Opportunity created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'

  /opportunities/{opportunityId}:
    get:
      tags:
        - Opportunities
      summary: Get an opportunity by ID
      operationId: getOpportunityById
      security:
        - bearerAuth: []
      parameters:
        - name: opportunityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested opportunity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
        '404':
          description: Opportunity not found.
    patch:
      tags:
        - Opportunities
      summary: Update an opportunity
      operationId: updateOpportunity
      security:
        - bearerAuth: []
      parameters:
        - name: opportunityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOpportunityDto'
      responses:
        '200':
          description: Opportunity updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
    delete:
      tags:
        - Opportunities
      summary: Delete an opportunity
      operationId: deleteOpportunity
      security:
        - bearerAuth: []
      parameters:
        - name: opportunityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Opportunity deleted successfully.

  /invoices:
    get:
      tags:
        - Invoices
      summary: List all invoices
      description: Retrieves a paginated list of invoices.
      operationId: getInvoices
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination.
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page.
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: A list of invoices.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
    post:
      tags:
        - Invoices
      summary: Create a new invoice
      description: Creates a new invoice for a lead.
      operationId: createInvoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceDto'
      responses:
        '201':
          description: Invoice created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          description: Invalid input data.

  /tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks
      operationId: getTasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      tags:
        - Tasks
      summary: Create a new task
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskDto'
      responses:
        '201':
          description: Task created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get a task by ID
      operationId: getTaskById
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found.
    patch:
      tags:
        - Tasks
      summary: Update a task
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskDto'
      responses:
        '200':
          description: Task updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      tags:
        - Tasks
      summary: Delete a task
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully.

components:
  schemas:
    CrmUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Alex Doe'
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, company-admin, super-admin]
        avatarUrl:
          type: string
          format: uri
          example: 'https://picsum.photos/seed/alex/100/100'

    Account:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          example: 'Innovate Inc.'
        industry:
          type: string
          example: 'Technology'
        ownerId:
          type: integer
        organizationId:
          type: integer
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        customFields:
          type: object
          additionalProperties: true
          description: "Flexible key-value store for custom data."
          example:
            { "sla_level": "Gold", "renewal_date": "2025-12-31" }

    CreateAccountDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          example: 'New Horizon Corp'
        industry:
          type: string
          example: 'Renewable Energy'
        ownerId:
          type: integer
          description: "ID of the user who owns this account."
          example: 1
        customFields:
          type: object
          additionalProperties: true
    
    Contact:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        accountId:
          type: integer
        organizationId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateContactDto:
      type: object
      required:
        - name
        - accountId
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        accountId:
          type: integer

    UpdateContactDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        accountId:
          type: integer

    Lead:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        email:
          type: string
          format: email
        source:
          type: string
        status:
          type: string
          enum: [New, Contacted, Qualified, Lost]
        ownerId:
          type: integer
        organizationId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateLeadDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        source:
          type: string
        status:
          type: string
          enum: [New, Contacted, Qualified, Lost]
        ownerId:
          type: integer

    UpdateLeadDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        source:
          type: string
        status:
          type: string
          enum: [New, Contacted, Qualified, Lost]
        ownerId:
          type: integer

    Opportunity:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        accountId:
          type: integer
        stage:
          type: string
          enum: [Prospecting, Qualification, Proposal, Closing, Won, Lost]
        amount:
          type: integer
        closeDate:
          type: string
          format: date
        ownerId:
          type: integer
        organizationId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateOpportunityDto:
      type: object
      required:
        - name
        - accountId
      properties:
        name:
          type: string
        accountId:
          type: integer
        stage:
          type: string
          enum: [Prospecting, Qualification, Proposal, Closing, Won, Lost]
        amount:
          type: integer
        closeDate:
          type: string
          format: date
        ownerId:
          type: integer
    
    UpdateOpportunityDto:
      type: object
      properties:
        name:
          type: string
        accountId:
          type: integer
        stage:
          type: string
          enum: [Prospecting, Qualification, Proposal, Closing, Won, Lost]
        amount:
          type: integer
        closeDate:
          type: string
          format: date
        ownerId:
          type: integer

    LoginDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    RegisterDto:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    Invoice:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        invoiceNumber:
          type: string
        leadId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [Draft, Sent, Paid, Void]
        dueDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time

    CreateInvoiceDto:
      type: object
      required:
        - leadId
        - amount
        - dueDate
      properties:
        leadId:
          type: integer
        amount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [Draft, Sent]
          default: Draft
    
    UpdateInvoiceDto:
      type: object
      properties:
        status:
          type: string
          enum: [Draft, Sent, Paid, Void]
    
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [Pending, Completed]
        assignedToId:
          type: integer
        relatedToType:
          type: string
          enum: [Account, Contact, Opportunity, Lead]
        relatedToId:
          type: integer
        organizationId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateTaskDto:
      type: object
      required:
        - title
        - dueDate
        - assignedToId
      properties:
        title:
          type: string
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [Pending, Completed]
        assignedToId:
          type: integer
        relatedToType:
          type: string
          enum: [Account, Contact, Opportunity, Lead]
        relatedToId:
          type: integer

    UpdateTaskDto:
      type: object
      properties:
        title:
          type: string
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [Pending, Completed]
        assignedToId:
          type: integer
        relatedToType:
          type: string
          enum: [Account, Contact, Opportunity, Lead]
        relatedToId:
          type: integer


  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []